Refactored refresh metric and get metric analysis to services
